+++ SUD/Models/SaleDetailBk.cs
@@ -34,12 +34,12 @@ namespace SUD.Models
         [Required(ErrorMessage = "Cantidad Requerida")]
         [Display(Name = "Porcentaje de IVA", Description = "0.00")]
         [RegularExpression(@"^[0-9]\d*(\.\d +)?$", ErrorMessage = "Porcentaje Invalido, solo se permiten numeros enteros.")]
-        public float IVAPercentage { get; set; }
+        public decimal IVAPercentage { get; set; }
 
         [Required(ErrorMessage = "Cantidad Requerida")]
         [Display(Name = "Porcentaje de descuento", Description = "0.00")]
         [RegularExpression(@"^[0-9]\d*(\.\d +)?$", ErrorMessage = "Porcentaje Invalido, solo se permiten numeros enteros.")]
-        public float DiscountRate { get; set; }
+        public decimal DiscountRate { get; set; }
 
         [Required]
         [Display(Name = "Producto")]
